backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pyjwt==2.8.0
werkzeug==2.3.7
python-dotenv==1.0.0
flask-mail==0.9.1

psycopg2-binary
